Release/release (#22)
* bug fix typo config

* add mixin schema

* enhance makefile

* enhance make clean

* implement version on model

* revert back trx on repo layer, tx only managed on service layer for easy maintenance

* add all schema using mixin as base schema field

* update latest echo version, move package, use echo logger

* fixing make clean

* modify mock

* standardize mock names

* bugfix user update transaction logic

* fixing logic user repo and service - blcoked by unit test

* completing unit test on utils package

* fixing unit test on user

* fixing unit test on user

+++ internal/applications/user/dto/user_dto.go
@@ -4,7 +4,7 @@ import "time"
 
 type (
 	UserRequest struct {
-		RoleId   int32  `json:"role_id" validate:"required"`
+		RoleId   uint64 `json:"role_id" validate:"required"`
 		Name     string `json:"name" validate:"required"`
 		Email    string `json:"email" validate:"required,email"`
 		Password string `json:"password" validate:"required,password"`
